+++ rsass/src/parser/formalargs.rs
@@ -4,12 +4,14 @@ use super::value::space_list;
 use super::{PResult, Span};
 use crate::sass::{CallArgs, FormalArgs, Name};
 use nom::bytes::complete::tag;
-use nom::combinator::{map, map_res, opt};
+use nom::character::complete::char;
+use nom::combinator::{cut, map, map_res, opt};
+use nom::error::context;
 use nom::multi::separated_list0;
 use nom::sequence::{delimited, pair, preceded, terminated};
 
 pub fn formal_args(input: Span) -> PResult<FormalArgs> {
-    let (input, _) = terminated(tag("("), opt_spacelike)(input)?;
+    let (input, _) = terminated(char('('), opt_spacelike)(input)?;
     let (input, v) = separated_list0(
         preceded(tag(","), opt_spacelike),
         map(
@@ -17,7 +19,7 @@ pub fn formal_args(input: Span) -> PResult<FormalArgs> {
                 delimited(tag("$"), name, opt_spacelike),
                 opt(delimited(
                     terminated(tag(":"), opt_spacelike),
-                    space_list,
+                    cut(context("Expected expression.", space_list)),
                     opt_spacelike,
                 )),
             ),
@@ -26,7 +28,7 @@ pub fn formal_args(input: Span) -> PResult<FormalArgs> {
     )(input)?;
     let (input, _) = terminated(opt(tag(",")), opt_spacelike)(input)?;
     let (input, va) = terminated(opt(tag("...")), opt_spacelike)(input)?;
-    let (input, _) = tag(")")(input)?;
+    let (input, _) = char(')')(input)?;
     Ok((
         input,
         if va.is_none() {
@@ -39,28 +41,31 @@ pub fn formal_args(input: Span) -> PResult<FormalArgs> {
 
 pub fn call_args(input: Span) -> PResult<CallArgs> {
     delimited(
-        terminated(tag("("), opt_spacelike),
+        terminated(char('('), opt_spacelike),
         map_res(
             pair(
                 separated_list0(
                     terminated(tag(","), opt_spacelike),
                     pair(
-                        opt(delimited(
-                            tag("$"),
-                            map(name, Name::from),
+                        opt(map(
                             delimited(
-                                ignore_comments,
-                                tag(":"),
-                                opt_spacelike,
+                                tag("$"),
+                                name,
+                                delimited(
+                                    ignore_comments,
+                                    char(':'),
+                                    opt_spacelike,
+                                ),
                             ),
+                            Name::from,
                         )),
                         terminated(space_list, opt_spacelike),
                     ),
                 ),
-                opt(terminated(tag(","), opt_spacelike)),
+                opt(terminated(char(','), opt_spacelike)),
             ),
             |(args, trail)| CallArgs::new(args, trail.is_some()),
         ),
-        tag(")"),
+        cut(char(')')),
     )(input)
 }