+++ rsass/tests/spec/directives/test_use/error/member/inaccessible.rs
@@ -64,6 +64,7 @@ mod private {
     }
 
     #[test]
+    #[ignore] // wrong result
     fn function() {
         let runner = runner().with_cwd("function");
         assert_eq!(