+++ rsass/tests/spec/directives/test_use/error/syntax/empty.rs
@@ -6,7 +6,6 @@ fn runner() -> crate::TestRunner {
 }
 
 #[test]
-#[ignore] // wrong error
 fn test() {
     assert_eq!(
         runner().err("@use;\n"),