+++ rsass/tests/spec/non_conformant/errors/fn_varargs/with_default.rs
@@ -6,7 +6,6 @@ fn runner() -> crate::TestRunner {
 }
 
 #[test]
-#[ignore] // wrong error
 fn test() {
     assert_eq!(
         runner().err("@function test($param...:\"default\") {}"),